Project_6_Fashion_MNIST_CNN_Classifier/app/requirements.txt
numpy==1.23.5
tensorflow==2.10.0
pandas==1.5.3
matplotlib==3.7.5
scikit-learn==1.2.2
seaborn==0.12.2
opencv-python==4.9.0.80
scipy==1.10.1
pillow==9.5.0
streamlit
jupyterlab

# streamlit>=1.24